refactor(Func): move database tables to models
Moved the RAMSTK database tables related to the function workstream to
the models package per the new API.

* Update set_attributes() method to new method topology.
* Remove calculation methods from model files.
* Move tests to tests/models/programdb.
* Update test suite to reflect changes.
