+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Refactor \_on\_combo\_changed\(\) and \_on\_focus\_out\(\) methods to eliminate duplicate code [\#200](https://github.com/ReliaQualAssociates/ramstk/issues/200) [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)]
 - PyGTK deprecation [\#58](https://github.com/ReliaQualAssociates/ramstk/issues/58)
 - Support Python 3 [\#57](https://github.com/ReliaQualAssociates/ramstk/issues/57)
+- update documentation [\#255](https://github.com/ReliaQualAssociates/ramstk/pull/255) ([weibullguy](https://github.com/weibullguy))
 - feature/revision refactor [\#254](https://github.com/ReliaQualAssociates/ramstk/pull/254) [[module:failuredefinition](https://github.com/ReliaQualAssociates/ramstk/labels/module:failuredefinition)] [[module:revision](https://github.com/ReliaQualAssociates/ramstk/labels/module:revision)] [[module:usage](https://github.com/ReliaQualAssociates/ramstk/labels/module:usage)] ([weibullguy](https://github.com/weibullguy))
 - Feature/make analyses mvc [\#253](https://github.com/ReliaQualAssociates/ramstk/pull/253) [[module:allocation](https://github.com/ReliaQualAssociates/ramstk/labels/module:allocation)] [[module:hardware](https://github.com/ReliaQualAssociates/ramstk/labels/module:hardware)] [[module:similaritem](https://github.com/ReliaQualAssociates/ramstk/labels/module:similaritem)] ([weibullguy](https://github.com/weibullguy))
 - Add target to docs/Makefile to auto publish gh-pages [\#223](https://github.com/ReliaQualAssociates/ramstk/pull/223) ([weibullguy](https://github.com/weibullguy))
