+++ tests/models/programdb/test_ramstkaction.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 #
-#       tests.dao.programdb.test_ramstkaction.py is part of The RAMSTK Project
+#       tests.models.programdb.test_ramstkaction.py is part of The RAMSTK
+#       Project
 #
 # All rights reserved.
+# Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """Test class for testing the RAMSTKAction module algorithms and models."""
 
 # Standard Library Imports
@@ -12,12 +14,7 @@ from datetime import date, timedelta
 import pytest
 
 # RAMSTK Package Imports
-from ramstk.dao.programdb.RAMSTKAction import RAMSTKAction
-
-__author__ = 'Doyle Rowland'
-__email__ = 'doyle.rowland@reliaqual.com'
-__organization__ = 'ReliaQual Associates, LLC'
-__copyright__ = 'Copyright 2017 Doyle "weibullguy" Rowland'
+from ramstk.models.programdb.RAMSTKAction import RAMSTKAction
 
 ATTRIBUTES = {
     'action_due_date': date.today() + timedelta(days=30),
@@ -29,21 +26,14 @@ ATTRIBUTES = {
     'action_recommended': b'Recommended action for Failure Cause #1',
     'action_category': '',
     'action_owner': '',
-    'cause_id': 1,
-    'action_id': 1,
     'action_approved': 0,
 }
 
 
 @pytest.mark.integration
 def test_ramstkaction_create(test_dao):
-    """
-    __init__() should create an RAMSTKAction model.
-    """
-    _session = test_dao.RAMSTK_SESSION(
-        bind=test_dao.engine, autoflush=False, expire_on_commit=False,
-    )
-    DUT = _session.query(RAMSTKAction).first()
+    """__init__() should create an RAMSTKAction model."""
+    DUT = test_dao.session.query(RAMSTKAction).first()
     assert isinstance(DUT, RAMSTKAction)
 
     # Verify class attributes are properly initialized.
@@ -69,10 +59,7 @@ def test_get_attributes(test_dao):
     """
     get_attributes() should return a dict of attribute:value pairs.
     """
-    _session = test_dao.RAMSTK_SESSION(
-        bind=test_dao.engine, autoflush=False, expire_on_commit=False,
-    )
-    DUT = _session.query(RAMSTKAction).first()
+    DUT = test_dao.session.query(RAMSTKAction).first()
 
     _attributes = DUT.get_attributes()
 
@@ -84,37 +71,26 @@ def test_set_attributes(test_dao):
     """
     set_attributes() should return a zero error code on success.
     """
-    _session = test_dao.RAMSTK_SESSION(
-        bind=test_dao.engine, autoflush=False, expire_on_commit=False,
-    )
-    DUT = _session.query(RAMSTKAction).first()
+    DUT = test_dao.session.query(RAMSTKAction).first()
 
-    _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
-
-    assert _error_code == 0
-    assert _msg == (
-        "RAMSTK SUCCESS: Updating RAMSTKAction {0:d} "
-        "attributes.".format(DUT.action_id)
-    )
+    assert DUT.set_attributes(ATTRIBUTES) is None
 
 
 @pytest.mark.integration
-def test_set_attributes_missing_key(test_dao):
-    """
-    set_attributes() should return a 40 error code when passed a dict with a missing key.
-    """
-    _session = test_dao.RAMSTK_SESSION(
-        bind=test_dao.engine, autoflush=False, expire_on_commit=False,
-    )
-    DUT = _session.query(RAMSTKAction).first()
+def test_set_attributes_none_value(test_dao):
+    """set_attributes() should set an attribute to it's default value when the attribute is passed with a None value."""
+    DUT = test_dao.session.query(RAMSTKAction).first()
 
-    ATTRIBUTES.pop('action_taken')
-    _error_code, _msg = DUT.set_attributes(ATTRIBUTES)
+    ATTRIBUTES['action_status'] = None
 
-    assert _error_code == 40
-    assert _msg == (
-        "RAMSTK ERROR: Missing attribute 'action_taken' in attribute "
-        "dictionary passed to RAMSTKAction.set_attributes()."
-    )
+    assert DUT.set_attributes(ATTRIBUTES) is None
+    assert DUT.get_attributes()['action_status'] == ''
 
-    ATTRIBUTES['action_taken'] = ''
+
+@pytest.mark.integration
+def test_set_attributes_unknown_attributes(test_dao):
+    """set_attributes() should raise an AttributeError when passed an unknown attribute."""
+    DUT = test_dao.session.query(RAMSTKAction).first()
+
+    with pytest.raises(AttributeError):
+        DUT.set_attributes({'shibboly-bibbly-boo': 0.9998})
